Be resilient to network failures
Let the main program run no matter what the connection status to the
backend is. Keep reconnecting in background if connection closes for
whatever reason (e.g. broken pipe).

As a side effect, first connection is created after the program is
started. This means we might miss initial part of output. But for the
user this means her program will run as usual without additional delays
imposed by rdoctor. There's a buffer for 100 lines which should cover
the standard scenario when backend is OK.

Also exit with program's exit code.